pyproject.toml
[project]
name = "openpipixia"
version = "0.1.0"
description = "A lightweight skills-first agent built with Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.0",
    "greenlet>=3.1.1",
    "lark-oapi>=1.4.15",
    "loguru>=0.7.3",
    "python-socks>=2.5.3",
    "websockets>=12.0",
    "litellm>=1.75.5,<1.80.17",
    "openai>=1.100.2",
    "httpx>=0.27.0",
    "oauth-cli-kit>=0.1.1",
]

[project.scripts]
openpipixia = "openpipixia.app.cli:main"
openpipixia-gui-mcp = "openpipixia.gui.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openpipixia"]

[tool.hatch.build]
include = [
    "openpipixia/**/*.py",
    "openpipixia/bridge/**",
    "openpipixia/skills/**",
]
